chore: migrate `Stage` components to `game.save` package and refactor model to use `record`
Moved `Stage` domain model, request/response mappers, persistence layer, and associated services to the `game.save` package. Updated `Stage` implementation from a class to a `record`. Refactored imports, service methods, and test cases for consistency.
