+++ lsadf_core/src/main/java/com/lsadf/core/domain/game/save/stage/validation/StageConsistency.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.core.domain.game.stage.validation;
+package com.lsadf.core.domain.game.save.stage.validation;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
